// packages/sdk-next/src/pool/offlinePoolService/types.ts
export enum AssetType {
  Bond = 'Bond',
  External = 'External',
  StableSwap = 'StableSwap',
  Token = 'Token',
  XYK = 'XYK',
  Erc20 = 'Erc20',
  PoolShare = 'PoolShare',
}

export type AssetDynamicFee = {
  assetFee: number;
  protocolFee: number;
  timestamp: number;
};

export interface PersistentAsset {
  id: string;
  decimals: number;
  symbol: string;
  existentialDeposit: string;
  isSufficient: boolean;
  type: AssetType;
  name?: string;
  icon?: string;
  location?: unknown;
  dynamicFee?: AssetDynamicFee;
}

export interface IPersistentPoolToken {
  id: string;
  decimals: number;
  symbol: string;
  balance: string;
  existentialDeposit: string;
  isSufficient: boolean;
  type: AssetType;
  tradable?: number;
  tradeable?: number;
  name?: string;
  icon?: string;
  location?: unknown;
  isWhiteListed?: boolean;
}
// ...
export interface IPersistentStableSwapBasePegSource {
  sourceKind: string;
  oracleName?: string;
  oraclePeriod?: string;
  oracleAsset?: string;
  valuePoints?: string[];
}

export interface IPersistentStableSwapBase {
  id: string;
  address: string;
  type: PoolType;
  tokens: IPersistentPoolToken[];
  maxInRatio: number;
  maxOutRatio: number;
  minTradingLimit: number;
  initialAmplification: number;
  finalAmplification: number;
  initialBlock: number;
  finalBlock: number;
  blockNumber: number;
  fee: number;
  totalIssuance: string;
  pegs: string[][];
  pegsUpdatedAtParaBlock?: number;
  maxPegUpdate?: number;
  pegSources?: IPersistentStableSwapBasePegSource[];
}
// ...
export type EmaOracleSource = string;
export type EmaOraclePeriod = string;

export interface IPersistentEmaOracleEntry {
  source: EmaOracleSource;
  period: EmaOraclePeriod;
  assets: string[];
  entry: IPersistentEmaOracleEntryData;
}

export interface IPersistentEmaOracleEntryData {
  price: IPersistentEmaOracleEntryRatio;
  volume: IPersistentEmaOracleEntryVolume;
  liquidity: IPersistentEmaOracleEntryLiquidity;
  updatedAt: number;
}

export interface IPersistentEmaOracleEntryLiquidity {
  a: string;
  b: string;
}

export interface IPersistentEmaOracleEntryVolume {
  aIn: string;
  bOut: string;
  aOut: string;
  bIn: string;
}

export interface IPersistentEmaOracleEntryRatio {
  n: string;
  d: string;
}

export interface IPersistentMmOracleEntry {
  address: string;
  price: string;
  decimals: number;
  updatedAt: number;
}
// ...
export interface IPersistentMetaData {
  paraBlockNumber: number;
  paraBlockHash: string;
  relayBlockNumber: number;
}